# Parallel operation changes