This PR introduces two changes to parallel operations:
* it removes the unnecessary container pointer from `bitmapContainerKey` struct. This has a surprisingly significant impact on performance in the private benchmark[1] I'm using.
* it pools `[]container` slice passed to `orFunc` in `ParOr`. The measured perf. impact is not as significant as in the first change, but it's also positive

[1] I hope, hope to make some progress on #113 soon! :-)