@lemire 2018-01-16
Merging.

If you get around to it, I'd like to know why (you think) the container point hurts the performance so much. Is that a caching issue?