@lemire 2018-01-16
+1 on the progress regarding the benchmarking!