@maciej 2018-01-16
@lemire I think the answer might have little to do with CPU caches.

I've used Apple's Xcode Instrumentation tool to get traces for both old and new code. It runs a 100 iterations of a set expression on digital advertising data on top of roaring (the private benchmark I'm using).

new:
```
6249	6249.0ms    8.2%	181.0	 	       github.com/RoaringBitmap/roaring.ParOr
5943	5943.0ms    7.8%	1473.0	 	        github.com/RoaringBitmap/roaring.(*bitmapContainerHeap).Next
4450	4450.0ms    5.8%	146.0	 	         github.com/RoaringBitmap/roaring.(*bitmapContainerHeap).popIncrementing
4300	4300.0ms    5.6%	2129.0	 	          container/heap.Fix
2171	2171.0ms    2.8%	2021.0	 	           container/heap.down
150	150.0ms    0.1%	4.0	 	            github.com/RoaringBitmap/roaring.(*bitmapContainerHeap).Swap
146	146.0ms    0.1%	41.0	 	             runtime.writebarrierptr
```

old:
```
12242	12242.0ms    9.2%	450.0	 	       github.com/RoaringBitmap/roaring.ParOr
11729	11729.0ms    8.8%	2476.0	 	        github.com/RoaringBitmap/roaring.(*bitmapContainerHeap).PopNextContainers
8526	8526.0ms    6.4%	657.0	 	         github.com/RoaringBitmap/roaring.(*bitmapContainerHeap).PopIncrementing
7782	7782.0ms    5.8%	3239.0	 	          container/heap.Fix
4543	4543.0ms    3.4%	2658.0	 	           container/heap.down
1885	1885.0ms    1.4%	1127.0	 	            github.com/RoaringBitmap/roaring.(*bitmapContainerHeap).Swap
758	758.0ms    0.5%	166.0	 	             runtime.typedmemmove
592	592.0ms    0.4%	295.0	 	              runtime.bulkBarrierPreWrite
79	79.0ms    0.0%	13.0	 	          runtime.typedmemmove
66	66.0ms    0.0%	36.0	 	           runtime.bulkBarrierPreWrite
```

Now, comparing the traces Swap functions on `bitmapContainerHeap` it looks like it takes Go 10x long to move the structure containing `container` compared to the one that does not contain it.

` runtime.typedmemmove` seems to be taking a bulk of that time. That's all what I can tell looking at the traces. What `runtime.typedmemmove` does and why is it taking so much CPU time – I don't know.