@maciej 2018-01-16
There is surprisingly a lot happening in a one line function when you see it's [x86 assembly code](https://gist.github.com/maciej/089087384b8b4cc275dea86c9a58ffde).