@lemire 2018-01-16
I would expect the containers to be pointers to structs, wouldn't you? If so, then you had two pointers, effectively, an int and a 16-bit int. So... what? 4x8 = 32 bytes? Ok. So now you streamlined to int, 16-bit and one pointer. So... what? 3x8 = 24 bytes... it is not the difference of 8 bytes (or whatever) that would make much of a difference.

Maybe there is an impact on garbage collection?