@maciej 2018-01-16
From what I understand the Go GC design (and I hardly do) the write barriers are necessary for the GC to function: https://github.com/golang/proposal/blob/master/design/17503-eliminate-rescan.md