@lemire 2018-01-16
I'm annoyed that such a small difference can make a 10x perf difference and I don't have a clue as to why that is.