@maciej 2018-01-17
https://groups.google.com/forum/#!topic/golang-nuts/0ZlHR30p2EQ let's hope someone on _golang-nuts_ group will be kind enough to explain that